# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Library Swagger UI",
        "contact": {}
    },
    "basePath": "/",
    "paths": {
...
        "/admin/borrower/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This api create borrower",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BORROWER"
                ],
                "summary": "CREATE BORROWER",
                "parameters": [
                    {
                        "description": "Borrower",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.BorrowerCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/genproto.Borrower"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/borrower/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This api updates borrower",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BORROWER"
                ],
                "summary": "UPDATES BORROWER",
                "parameters": [
                    {
                        "description": "Borrower",
                        "name": "borrower",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.BorrowerCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "borrower updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/borrower/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This api deletes borrower by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BORROWER"
                ],
                "summary": "DELETES BORROWER",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BORROWER ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "borrower deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/borrower/all": {
            "get": {
                "description": "This api get all borrowers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BORROWER"
                ],
                "summary": "GET ALL BORROWERS",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.Borrowers"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/borrower/history/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This api GETS borrower by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BORROWER"
                ],
                "summary": "GET USER BORROWER HISTORY",
                "parameters": [
                    {
                        "type": "string",
                        "description": "USER ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.BorrowingHistory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/borrower/overdue": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This API retrieves a list of books that are overdue.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BORROWER"
                ],
                "summary": "Get Overdue Books",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.BorrowedBooks"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/borrower/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This api GETS borrower by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BORROWER"
                ],
                "summary": "GET USERS BORROWER",
                "parameters": [
                    {
                        "type": "string",
                        "description": "USER ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.BorrowedBooks"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/borrower/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This api get borrower by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BORROWER"
                ],
                "summary": "GET BORROWER",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BORROWER ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.Borrower"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "genproto.Book": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "genre_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "genproto.BorrowedBook": {
            "type": "object",
            "properties": {
                "book": {
                    "$ref": "#/definitions/genproto.Book"
                },
                "borrow_date": {
                    "type": "string"
                },
                "return_date": {
                    "type": "string"
                }
            }
        },
        "genproto.BorrowedBooks": {
            "type": "object",
            "properties": {
                "books": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.BorrowedBook"
                    }
                }
            }
        },
        "genproto.Borrower": {
            "type": "object",
            "properties": {
                "BookID": {
                    "type": "string"
                },
                "BorrowDate": {
                    "type": "string"
                },
                "Id": {
                    "type": "string"
                },
                "ReturnDate": {
                    "type": "string"
                },
                "UserID": {
                    "type": "string"
                }
            }
        },
        "genproto.BorrowerCreate": {
            "type": "object",
            "properties": {
                "BookID": {
                    "type": "string"
                },
                "BorrowDate": {
                    "type": "string"
                },
                "Id": {
                    "type": "string"
                },
                "ReturnDate": {
                    "type": "string"
                },
                "UserID": {
                    "type": "string"
                }
            }
        },
        "genproto.Borrowers": {
            "type": "object",
            "properties": {
                "Borrowers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.Borrower"
                    }
                }
            }
        },
        "genproto.BorrowingHistory": {
            "type": "object",
            "properties": {
                "books": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.BorrowedBook"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}